+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdUpdate.kt
@@ -11,6 +11,10 @@ import org.simple.clinic.patient.PatientPrefillInfo
 import org.simple.clinic.patient.businessid.Identifier
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.BpPassport
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.IndiaNationalHealthId
+import org.simple.clinic.patient.onlinelookup.api.LookupPatientOnline
+import org.simple.clinic.patient.onlinelookup.api.LookupPatientOnline.Result.Found
+import org.simple.clinic.patient.onlinelookup.api.LookupPatientOnline.Result.NotFound
+import org.simple.clinic.patient.onlinelookup.api.LookupPatientOnline.Result.OtherError
 import org.simple.clinic.platform.crash.CrashReporter
 import org.simple.clinic.scanid.EnteredCodeValidationResult.Failure
 import org.simple.clinic.scanid.EnteredCodeValidationResult.Success
@@ -35,6 +39,17 @@ class ScanSimpleIdUpdate @Inject constructor(
       is PatientSearchByIdentifierCompleted -> patientSearchByIdentifierCompleted(model, event)
       is ScannedQRCodeJsonParsed -> scannedQRCodeParsed(model, event)
       InvalidQrCode -> next(model.notSearching().invalidQrCode())
+      is OnlinePatientLookupWithIdentifierCompleted -> onlinePatientLookupWithIdentifierCompleted(model, event)
+    }
+  }
+
+  private fun onlinePatientLookupWithIdentifierCompleted(
+      model: ScanSimpleIdModel,
+      event: OnlinePatientLookupWithIdentifierCompleted
+  ): Next<ScanSimpleIdModel, ScanSimpleIdEffect> {
+    return when(event.result) {
+      is NotFound, is OtherError -> next(model.notSearching(), OpenPatientSearch(event.identifier, null, model.patientPrefillInfo))
+      else -> noChange() // update the list of patients or patient received here
     }
   }
 
@@ -72,13 +87,11 @@ class ScanSimpleIdUpdate @Inject constructor(
       model: ScanSimpleIdModel,
       event: PatientSearchByIdentifierCompleted
   ): Next<ScanSimpleIdModel, ScanSimpleIdEffect> {
-    val effect = if (event.patients.isEmpty()) {
-      OpenPatientSearch(event.identifier, null, model.patientPrefillInfo)
+    return if (event.patients.isEmpty()) {
+      dispatch(OnlinePatientLookupWithIdentifier(event.identifier))
     } else {
-      patientFoundByIdentifierSearch(patients = event.patients, identifier = event.identifier)
+      next(model = model.notSearching(), patientFoundByIdentifierSearch(patients = event.patients, identifier = event.identifier))
     }
-
-    return next(model = model.notSearching(), effect)
   }
 
   private fun patientFoundByIdentifierSearch(